[新增功能] 1.每个选项增加一个权值，根据权制计算最终结果； 2.根据model计算部分调整view实现；

+++ Classes/View/MtnnContentCell.h
@@ -20,6 +20,6 @@ typedef NS_ENUM(NSInteger, MtnnContentCellType)
 
 - (void)setCellType:(MtnnContentCellType)cellType;
 
-- (void)setCellWithName:(NSString *)name currentValue:(NSUInteger)currentValue allValue:(NSUInteger)allValue got:(BOOL)got;
+- (void)setCellWithName:(NSString *)name currentValue:(CGFloat)currentValue allValue:(CGFloat)allValue got:(BOOL)got;
 
 @end
